doc/requirements.txt
sphinx>=5.0
sphinx-rtd-theme
sphinxcontrib-mermaid
